requirements.txt
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.13.4
Brotli==1.1.0
cachetools==5.5.2
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
deep-translator==1.11.4
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.104.1
google-api-core==2.25.0
google-api-python-client==2.108.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.10
jeepney==0.9.0
langdetect==1.0.9
mutagen==1.47.0
openai==1.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pyparsing==3.2.3
python-dotenv==1.0.0
requests==2.32.3
rsa==4.9.1
SecretStorage==3.3.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.27.0
tqdm==4.67.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.24.0
websockets==15.0.1
yt-dlp==2025.5.22
